requirements.txt
python-telegram-bot>=22.5
git+https://github.com/pcider/bambulabs_api.git